prov/tcp: Fix resource leak from xnet_mr_regattr
Coverity states:
"The system resource will not be reclaimed and reused, reducing the
future availability of the resource.
In xnet_mplex_mr_regattr: Leak of memory or pointers to system
resources"

Fix: skip mr regattr if not in rma mode. See comment for more
information about future developement, such as FI_MR_HMEM support in
tcp.

Signed-off-by: Alex McKinley <alex.mckinley@intel.com>